Agregado metodo rating para eliminar, authentication user validator.

+++ Fullstack/001-fullstack/backend/src/main/java/com/bitebooking/controller/RatingController.java
@@ -1,8 +1,7 @@
 package com.bitebooking.controller;
 
-import com.bitebooking.model.Booking;
-import com.bitebooking.model.Menu;
-import com.bitebooking.model.Rating;
+import com.bitebooking.exception.UnauthorizedException;
+import com.bitebooking.model.*;
 import com.bitebooking.repository.BookingRepository;
 import com.bitebooking.repository.RatingRepository;
 import com.bitebooking.security.SecurityUtils;
@@ -12,6 +11,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @CrossOrigin("*")
 @RestController
@@ -23,20 +23,21 @@ public class RatingController {
     private final RatingRepository ratingRepository;
 
 
-
     @GetMapping("ratings/{id}")
     public Rating findById(@PathVariable Long id) {
 
         return this.ratingRepository.findById(id).orElseThrow();
     }
+
     @GetMapping("menus/filter-by-menu/{id}")
-    public List<Rating> findAllByMenuId(@PathVariable Long id){
+    public List<Rating> findAllByMenuId(@PathVariable Long id) {
         return this.ratingRepository.findByMenu_IdOrderByIdDesc(id);
     }
 
 
     @PostMapping("ratings")
     public Rating create(@RequestBody Rating rating) {
+
 
         SecurityUtils.getCurrentUser().ifPresent(user -> rating.setUser(user));
         return this.ratingRepository.save(rating);
@@ -52,9 +53,21 @@ public class RatingController {
     }
 
     @DeleteMapping("ratings/{id}")
-    public void deleteById(@PathVariable Long id){
+    public void deleteById(@PathVariable Long id) {
 
-        this.ratingRepository.deleteById(id);
+       Rating rating = this.ratingRepository.findById(id).orElseThrow();
+       User user = SecurityUtils.getCurrentUser().orElseThrow();
+
+        if(user.getRole().equals(Role.ADMIN) ||
+                (rating.getUser() != null && rating.getUser().getId().equals(user.getId()))
+        )
+            this.ratingRepository.deleteById(id);
+
+        else
+            throw new UnauthorizedException("No puede borrar el rating");
+
+
     }
+
 
 }
